+++ satella/coding/__init__.py
@@ -9,7 +9,8 @@ from .algos import merge_dicts
 from .recast_exceptions import rethrow_as, silence_excs
 from .typecheck import typed, Callable, Sequence, \
     TypeVar, Mapping, Iterable, Any, Optional, CallSignature, \
-    Number, coerce, Set, Dict, List, Tuple, checked_coerce, for_argument
+    Number, coerce, Set, Dict, List, Tuple, checked_coerce, for_argument, \
+    precondition, PreconditionError
 from .structures import TimeBasedHeap, Heap, typednamedtuple, OmniHashableMixin
 
 __all__ = [
@@ -19,7 +20,8 @@ __all__ = [
     'typed', 'NewType', 'Callable', 'Sequence', 'coerce'
     'TypeVar','Mapping', 'Iterable', 'Union', 'Any', 'Optional',
     'CallSignature', 'Number',
-    'Set', 'Dict', 'List', 'Tuple', 'checked_coerce', 'for_argument'
+    'Set', 'Dict', 'List', 'Tuple', 'checked_coerce', 'for_argument',
+    'precondition', 'PreconditionError',
     'rethrow_as', 'silence_excs'
 ]